[external assets] move api endpoint tests
